+++ apps/extension/src/ui/apps/dashboard/routes/Earn/DashboardEarnPage.tsx
@@ -112,7 +112,7 @@ const EarnPageHeader = () => {
   return (
     <div className="flex h-64 items-center rounded-[0.4688rem] border border-grey-800 px-6 py-8 text-left text-base text-body-secondary">
       <div className="flex w-full flex-col gap-4">
-        <div className="text-body-secondary text-sm">{t("Yield-Eligible Capital")}</div>
+        <div className="text-body-secondary text-sm">{t("Earn-Eligible Capital")}</div>
         <div className="font-bold text-2xl text-body">
           {!eligibleTotal && status === "loading" ? (
             <div className="animate-pulse rounded bg-grey-700 text-grey-700">$0.00</div>
