+++ apps/extension/src/ui/domains/Earn/seek/useSeekStakingModal.ts
@@ -1,5 +1,8 @@
 import { createGlobalOpenClose } from "@ui/hooks/createGlobalOpenClose"
 
+// these values double as contract function names: SeekStakingModal encodes the transaction with
+// `encodeFunctionData({ abi: seekSinglePoolStakingAbi, functionName: action })`, so each value
+// must match a function of that ABI — do not rename them for display purposes
 export type SeekStakingAction =
   | "stake"
   | "requestWithdrawal"
